+++ packages/insights-charts/source/js/charts/Chart.js
@@ -22,13 +22,12 @@ class Chart {
     }
   }
 
-  getPlotOptions(type, stackable = true) {
+  getPlotOptions(type, data, stackable = true) {
     const options = this.options[type] || {};
-    const series = this.data.getSeries();
 
-    if (series.length <= 1) {
+    if (data.length <= 1) {
       options.layout = 'normal';
-    } else if (series.length > 1 && !options.layout && stackable) {
+    } else if (data.length > 1 && !options.layout && stackable) {
       options.layout = 'stacked';
     }