+++ packages/insights-charts/source/js/charts/ScatterChart.js
@@ -63,7 +63,7 @@ class ScatterChart extends Chart {
         const yAxis = new YAxis(y, dimensions, yOptions, yAxisIndex);
         yAxis.render(svg);
 
-        const plotOptions = deepmerge(options, this.getPlotOptions(this.type));
+        const plotOptions = deepmerge(options, this.getPlotOptions(this.type, data));
 
         const seriesPoi = new SeriesPoi(
           data,
@@ -140,7 +140,7 @@ class ScatterChart extends Chart {
       const scale = this.yScales[yAxisIndex];
 
       if (scale) {
-        const plotOptions = deepmerge(options, this.getPlotOptions(this.type));
+        const plotOptions = deepmerge(options, this.getPlotOptions(this.type, data));
         const y = scale.yScale.update(data, yOptions, plotOptions.layout, dimensions, options);
 
         if (yAxisIndex === 0) {